[FIX] hr_contract_update_overtime: tests compatible with hr_holidays_attendance
If you run this module's tests while it's being installed for the 1st time, `hr_holidays_attendance` will enter its test scope and add extra constrains to which leaves are considered for employees. If we don't follow those constrains, the tests fail.

Thus here I'm using that module lazily, while not being strictly required as a dependency. If installed, I fulfill the data required for it to work, and then tests are more resilient.

@moduon MT-6583